+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes will be documented in this file.
 
 ### Changed
 
+* Report last traded as the aggregate of all fills (#84)
 * DropCopy doesn't need to subscribe positions (#79)
 * Now subscribing to position updates (FIX) (#79)